ComminWeath

Documentation for ComminWeath.

ComminWeath.Detrital

ComminWeath.Grain

ComminWeath.Rind

ComminWeath.WxAuth

ComminWeath.Tanaka2015

ComminWeath.TaylorI

ComminWeath.TaylorIII

ComminWeath.TaylorIV

ComminWeath.calcU

ComminWeath.calcslopes

ComminWeath.comminweath

ComminWeath.drawdate

ComminWeath.drawdates

ComminWeath.linreg

ComminWeath.Detrital — Type

Detrital(cU, r48, 408)

mutable struct for initial detrital grain U-series parameters.

Constructor function Detrital() builds an instance with default values.

field	units	description

cU	ng/g	concentration of U in detrital phase

r48		initial (²³⁴U/²³⁸U) of detrital grain

r08		initial (²³⁰Th/²³⁸U) of detrital grain

ComminWeath.Grain — Type

Grain(K, rfr, rho, L234, L230)

mutable struct for grain morphology and α-recoil parameters.

Constructor function Grain() builds an instance with default values.

field	units	description

K		grain shape factorⁱ

rfr		surface roughness factorⁱⁱ

rho	kg/m³	bulk density

L234	nm	recoil length of ²³⁴U

L230	nm	recoil length of ²³⁰Th

ⁱ after Cartwright, 1962, doi:10.1093/annhyg/5.3.163, ⁱⁱ after White et al., 1996, doi: 10.1016/0016-7037(96)00106-8

ComminWeath.Rind — Type

Rind(evolve, p, z, rho, cU, r48, 408)

mutable struct for authigenic soluble rind parameters.

Constructor function Rind() builds an instance with default values.

field	units	description

evolve	(boolean)	rind composition evolves (ON = true)

p		proportion of grain covered by rind

z	μm	thickness of rind

rho	kg/m³	bulk density of rind phase

cU	ng/g	concentration of U in rind phase

r48		initial (²³⁴U/²³⁸U) of rind phase

r08		initial (²³⁰Th/²³⁸U) of rind phase

ComminWeath.WxAuth — Type

WxAuth(sa_dependent, k, k_power, cU, r48, 408)

mutable struct for parameters of chemical weathering / authigenic replacement.

Constructor function WxAuth() builds an instance with default values.

field	units	description

sa_dependent	(boolean)	surface area dependence (ON = true)

k	g m⁻² a⁻¹	rate of detrital phase alteration to authigenic phaseⁱ

k_power		exponent on k for (time-dependent) power-law weathering rate

cU	ng/g	concentration of U in authigenic phase

r48		initial (²³⁴U/²³⁸U) of authigenic phase

r08		initial (²³⁰Th/²³⁸U) of authigenic phase

ⁱg m⁻² a⁻¹ if sa_exponent=true, (g g⁻¹) a⁻¹ if sa_exponent=false. This is an initial rate if k_power ≠ 1. ⁱⁱNot explored in this study, so it is set to 1 for a constant weathering rate, but it's there if you want it!

ComminWeath.Tanaka2015 — Method

ImpactChron.Tanaka2015()

Returns a NamedTuple of U-series data explicitly illustrating the effects of α-recoil implantation (Tanaka+ 2015, Chemical Geology, https://doi.org/10.1016/j.chemgeo.2014.12.025).

The data are from KOL48 (bulk) and magnetic susceptibility separates KOL48A–D (Table 3).

fields	description

r48	(²³⁴U/²³⁸U)

u48	2σ uncertainty of (²³⁴U/²³⁸U)

r48	(²³⁰Th/²³⁸U)

u48	2σ uncertainty of (²³⁰Th/²³⁸U)

ComminWeath.TaylorI — Method

TaylorI()

Returns a NamedTuple of U-series data from detrital fractions of Taylor I drift sediments (measured in this study).

fields	description

r48	(²³⁴U/²³⁸U)

u48	2σ uncertainty of (²³⁴U/²³⁸U)

r48	(²³⁰Th/²³⁸U)

u48	2σ uncertainty of (²³⁰Th/²³⁸U)

ComminWeath.TaylorIII — Method

TaylorIII()

Returns a NamedTuple of U-series data from detrital fractions of Taylor III drift sediments (measured in this study).

fields	description

r48	(²³⁴U/²³⁸U)

u48	2σ uncertainty of (²³⁴U/²³⁸U)

r48	(²³⁰Th/²³⁸U)

u48	2σ uncertainty of (²³⁰Th/²³⁸U)

ComminWeath.TaylorIV — Method

TaylorIV()

Returns a NamedTuple of U-series data from detrital fractions of Taylor IV drift sediments (measured in this study).

fields	description

r48	(²³⁴U/²³⁸U)

u48	2σ uncertainty of (²³⁴U/²³⁸U)

r48	(²³⁰Th/²³⁸U)

u48	2σ uncertainty of (²³⁰Th/²³⁸U)

ComminWeath.calcU — Method

calcU(a, t; cU, g=Grain(), d=Detrital(), wx=WxAuth(),r=Rind())

Calculate the U abundance for sediments with grain sizes in a at times in t (ka).

A Vector of U concentrations (ng/g) may be passed to cU to simulate grain-size dependent U concentrations, but this requires length(cU) == length(a).

Custom parameterizations for Grain, Detrital, WxAuth, and Rind may be passed to g, d, wx and r, respectively.

see also: Grain, Detrital, WxAuth, Rind

ComminWeath.calcslopes — Method

calcslopes(t; a, cU, g=Grain(), d=Detrital(), wx=WxAuth(), r=Rind())

Calculate the (²³⁴U/²³⁸U)-(²³⁰Th/²³⁸U) slopes of grain sizes in a (default= [10.,20.,30.,40.]) at times given in t (ka).

A Vector of U concentrations (ng/g) may be passed to cU to simulate grain-size dependent U concentrations, but this requires length(cU) == length(a).

Custom parameterizations for Grain, Detrital, WxAuth, and Rind may be passed to g, d, wx and r, respectively.

see also: Grain, Detrital, WxAuth, Rind

ComminWeath.comminweath — Method

function comminweath(diameter, grain, detrital, wxauth, rind; timeseries)

Calculates post-comminution U-series evolutions for a grain of a given diameter (in μm) over a timeseries (<: AbstractRange) in years (default=0:1:2e6).

The history depends on grain morphology as well as the detrital, authigenic weathering (wxauth), and soluble authigenic rind parameters; respectively represented by mutable struct types Grain, Detrital, WxAuth, and Rind. See the docs of these types for more details.

Outputs a NamedTuple containing the timeseries and vectors of U elemental and isotopic evolutions with corresponding indices.

key	description

t	timeseries

A234	(²³⁴U/²³⁸U) of insoluble (detrital) grain

A230	(²³⁰Th/²³⁸U) of insoluble (detrital) grain

cU	[U] of insoluble (detrital) grain

etc	NamedTuple wrapping rind_ and bulk_ isotopic and [U] evolutions.

ComminWeath.drawdate — Method

drawdates(dates, evs::NamedTuple)

Returns the (²³⁴U/²³⁸U) (A234), (²³⁰Th/²³⁸U) (A230), and U concentration (cU) of a sediment grain for a dates (in ka), given a U-series history (evs) from the comminweath function.

ComminWeath.drawdates — Method

drawdates(dates, evs::NamedTuple)

Returns the (²³⁴U/²³⁸U) (A234), (²³⁰Th/²³⁸U) (A230), and U concentration (cU) of a sediment grain for a collection of dates (in ka), given a U-series history (evs) from the comminweath function.

ComminWeath.linreg — Method

linreg(x::Vector,y::Vector)

Calculate a linear regression by QR decomposition for values of independent variable x and dependent variable y.

Returns a NamedTuple with slope m, intercept b, and coefficient of determination r²